[Agent] extract location display service
