+++ tests/unit/services/entityDisplayDataProvider.test.js
@@ -7,15 +7,14 @@ import {
   PORTRAIT_COMPONENT_ID,
   DESCRIPTION_COMPONENT_ID,
   POSITION_COMPONENT_ID,
-  EXITS_COMPONENT_ID,
 } from '../../../src/constants/componentIds.js';
-import { CORE_MOD_ID } from '../../../src/constants/core';
 import { LOGGER_INFO_METHOD_ERROR } from '../../common/constants.js';
 
 describe('EntityDisplayDataProvider', () => {
   let mockEntityManager;
   let mockLogger;
   let mockSafeEventDispatcher;
+  let mockLocationDisplayService;
   let service;
 
   beforeEach(() => {
@@ -31,10 +30,15 @@ describe('EntityDisplayDataProvider', () => {
     mockSafeEventDispatcher = {
       dispatch: jest.fn(),
     };
+    mockLocationDisplayService = {
+      getLocationDetails: jest.fn(),
+      getLocationPortraitData: jest.fn(),
+    };
     service = new EntityDisplayDataProvider({
       entityManager: mockEntityManager,
       logger: mockLogger,
       safeEventDispatcher: mockSafeEventDispatcher,
+      locationDisplayService: mockLocationDisplayService,
     });
   });
 
@@ -57,6 +61,28 @@ describe('EntityDisplayDataProvider', () => {
           })
       ).toThrow(
         "Invalid or missing method 'getEntityInstance' on dependency 'entityManager'."
+      );
+    });
+
+    it('should throw an error if locationDisplayService is missing or invalid', () => {
+      expect(
+        () =>
+          new EntityDisplayDataProvider({
+            entityManager: mockEntityManager,
+            logger: mockLogger,
+            safeEventDispatcher: mockSafeEventDispatcher,
+          })
+      ).toThrow('Missing required dependency: locationDisplayService.');
+      expect(
+        () =>
+          new EntityDisplayDataProvider({
+            entityManager: mockEntityManager,
+            logger: mockLogger,
+            safeEventDispatcher: mockSafeEventDispatcher,
+            locationDisplayService: {},
+          })
+      ).toThrow(
+        "Invalid or missing method 'getLocationDetails' on dependency 'locationDisplayService'."
       );
     });
 
@@ -432,168 +458,32 @@ describe('EntityDisplayDataProvider', () => {
         expect.stringContaining('called with null or empty entityId')
       );
     });
-  });
 
-  // getLocationDetails
-  describe('getLocationDetails', () => {
-    it('should return compiled location details with exits', () => {
-      const mockLocationEntity = {
-        id: 'loc1',
-        getComponentData: jest.fn((componentId) => {
-          if (componentId === NAME_COMPONENT_ID) return { text: 'Grand Hall' };
-          if (componentId === DESCRIPTION_COMPONENT_ID)
-            return { text: 'A vast hall.' };
-          if (componentId === EXITS_COMPONENT_ID)
-            return [
-              { direction: 'north', target: 'loc2' },
-              { direction: 'south', target: 'loc3' },
-              { direction: 'a secret passage', target: 'loc_secret' },
-              { target: 'loc_no_dir' },
-              null,
-              { direction: '  ', target: 'loc_empty_dir' },
-            ];
-          return null;
-        }),
-      };
-      mockEntityManager.getEntityInstance.mockReturnValue(mockLocationEntity);
-      // ** FIX: Remove spies on service.getEntityName and service.getEntityDescription **
-      // Allow the actual methods to run and call mockLocationEntity.getComponentData
-
-      const details = service.getLocationDetails('loc1');
-      expect(details).toEqual({
-        name: 'Grand Hall',
-        description: 'A vast hall.',
-        exits: [
-          { description: 'north', target: 'loc2', id: 'loc2' },
-          { description: 'south', target: 'loc3', id: 'loc3' },
-          {
-            description: 'a secret passage',
-            target: 'loc_secret',
-            id: 'loc_secret',
-          },
-          {
-            description: 'Unspecified Exit',
-            target: 'loc_no_dir',
-            id: 'loc_no_dir',
-          },
-          {
-            description: 'Unspecified Exit',
-            target: 'loc_empty_dir',
-            id: 'loc_empty_dir',
-          },
-        ],
+    describe('location-based delegates', () => {
+      it('getLocationDetails delegates to service', () => {
+        mockLocationDisplayService.getLocationDetails.mockReturnValue({
+          name: 'n',
+          description: 'd',
+          exits: [],
+        });
+        const result = service.getLocationDetails('loc1');
+        expect(
+          mockLocationDisplayService.getLocationDetails
+        ).toHaveBeenCalledWith('loc1');
+        expect(result).toEqual({ name: 'n', description: 'd', exits: [] });
       });
-      expect(mockEntityManager.getEntityInstance).toHaveBeenCalledWith('loc1');
-      // Now these internal calls should happen:
-      expect(mockLocationEntity.getComponentData).toHaveBeenCalledWith(
-        NAME_COMPONENT_ID
-      );
-      expect(mockLocationEntity.getComponentData).toHaveBeenCalledWith(
-        DESCRIPTION_COMPONENT_ID
-      );
-      expect(mockLocationEntity.getComponentData).toHaveBeenCalledWith(
-        EXITS_COMPONENT_ID
-      );
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        expect.stringContaining(
-          "Invalid exit item in exits component for location 'loc1'"
-        ),
-        expect.any(Object)
-      );
-    });
 
-    it('should return null if location entity not found', () => {
-      mockEntityManager.getEntityInstance.mockReturnValue(null);
-      expect(service.getLocationDetails('nonExistentLoc')).toBeNull();
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        expect.stringContaining(
-          "Location entity with ID 'nonExistentLoc' not found."
-        )
-      );
-    });
-
-    it('should handle EXITS_COMPONENT_ID data being present but not an array', () => {
-      const mockLocationEntity = {
-        id: 'loc_bad_exits',
-        getComponentData: jest.fn((id) => {
-          if (id === EXITS_COMPONENT_ID) return { not_an_array: 'bad_data' };
-          return null;
-        }),
-      };
-      mockEntityManager.getEntityInstance.mockReturnValue(mockLocationEntity);
-
-      service.getLocationDetails('loc_bad_exits');
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        expect.stringContaining(
-          "Exits component data for location 'loc_bad_exits' is present but not an array."
-        ),
-        expect.any(Object)
-      );
-    });
-
-    it('should return null if locationEntityId is null or empty', () => {
-      expect(service.getLocationDetails(null)).toBeNull();
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        expect.stringContaining('called with null or empty locationEntityId')
-      );
-    });
-  });
-
-  // getLocationPortraitData
-  describe('getLocationPortraitData', () => {
-    it('should return portrait data with alt text', () => {
-      const mockEntity = {
-        id: 'loc1',
-        definitionId: 'core:loc',
-        getComponentData: jest.fn().mockReturnValue({
-          imagePath: 'images/loc.png',
-          altText: 'Location Alt',
-        }),
-      };
-      mockEntityManager.getEntityInstance.mockReturnValue(mockEntity);
-
-      const result = service.getLocationPortraitData('loc1');
-
-      expect(result).toEqual({
-        imagePath: '/data/mods/core/images/loc.png',
-        altText: 'Location Alt',
+      it('getLocationPortraitData delegates to service', () => {
+        mockLocationDisplayService.getLocationPortraitData.mockReturnValue({
+          imagePath: '/path/img.png',
+          altText: 'alt',
+        });
+        const result = service.getLocationPortraitData('loc1');
+        expect(
+          mockLocationDisplayService.getLocationPortraitData
+        ).toHaveBeenCalledWith('loc1');
+        expect(result).toEqual({ imagePath: '/path/img.png', altText: 'alt' });
       });
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        expect.stringContaining(
-          "Constructed portrait path for location 'loc1': /data/mods/core/images/loc.png"
-        )
-      );
-    });
-
-    it('should return null if portrait component missing', () => {
-      const mockEntity = {
-        id: 'loc2',
-        definitionId: 'core:loc',
-        getComponentData: jest.fn().mockReturnValue(null),
-      };
-      mockEntityManager.getEntityInstance.mockReturnValue(mockEntity);
-      const result = service.getLocationPortraitData('loc2');
-      expect(result).toBeNull();
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        expect.stringContaining(
-          "Location entity 'loc2' has no valid PORTRAIT_COMPONENT_ID data or imagePath."
-        )
-      );
-    });
-
-    it('should return null if entity not found', () => {
-      mockEntityManager.getEntityInstance.mockReturnValue(null);
-      expect(service.getLocationPortraitData('missing')).toBeNull();
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        expect.stringContaining("Location entity with ID 'missing' not found.")
-      );
-    });
-
-    it('should return null if locationEntityId is null or empty', () => {
-      expect(service.getLocationPortraitData(null)).toBeNull();
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        expect.stringContaining('called with null or empty locationEntityId')
-      );
     });
   });
 });
